aligned login and register view to fit MVVM model
