+++ View/LoginWindow.xaml
@@ -62,7 +62,7 @@
                     </Style>
                 </Button.Resources>
             </Button>
-            <Label Name="lblForgotPass" MouseDoubleClick="lblForgotPassDoubleClick" Grid.Row="4" Grid.Column="1" HorizontalAlignment="Right" VerticalAlignment="Center" FontSize="15" Content="Forgot Password?" Foreground="#EB5E28" FontFamily="Consolas" Cursor="Hand"/>
+            <Label Name="lblForgotPass" Grid.Row="4" Grid.Column="1" HorizontalAlignment="Right" VerticalAlignment="Center" FontSize="15" Content="Forgot Password?" Foreground="#EB5E28" FontFamily="Consolas" Cursor="Hand"/>
             <Border Grid.Row="6" Grid.Column="1" Width="Auto" VerticalAlignment="Top" BorderThickness="1,1,1,1" BorderBrush="#CCC5B9"/>
             <Grid Grid.Row="6" Grid.Column="1">
                 <Grid.ColumnDefinitions>
